Fixed and change folder title component
