+++ src/Components/About/About.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Title from './Title';
+import Title from '../Title';
 
 const About = () => {
   const [text, setText] = useState('');
