+++ src/Components/About/index.js
@@ -1,0 +1,2 @@
+export { default } from "./About";
+export * from "./About";
