+++ src/Components/Title/index.js
@@ -1,0 +1,2 @@
+export { default } from "./Title";
+export * from "./Title";
